Revamp update-versions workflow (#7868)
