+++ hack/versions/requirements.txt
@@ -1,0 +1,2 @@
+rich>=13
+ruamel.yaml==0.17.31
